+++ frontend/src/app/[locale]/subscribe/SubscriptionForm.tsx
