+++ frontend/src/app/[locale]/subscribe/confirmation/page.tsx
