+++ frontend/src/app/[locale]/subscribe/page.tsx
@@ -1,6 +1,6 @@
 import pick from "lodash/pick";
 import { Metadata } from "next";
-import SubscriptionForm from "src/app/[locale]/newsletter/NewsletterForm";
+import SubscriptionForm from "src/app/[locale]/subscribe/SubscriptionForm";
 import { SUBSCRIBE_CRUMBS } from "src/constants/breadcrumbs";
 
 import {
