+++ frontend/src/constants/breadcrumbs.ts
@@ -3,18 +3,18 @@ import { Breadcrumb, BreadcrumbList } from "src/components/Breadcrumbs";
 const HOME: Breadcrumb = { title: "Home", path: "/" };
 const RESEARCH: Breadcrumb = { title: "Research", path: "/research/" };
 const PROCESS: Breadcrumb = { title: "Process", path: "/process/" };
-const SUBSCRIBE: Breadcrumb = { title: "Subscribe", path: "/newsletter/" };
+const SUBSCRIBE: Breadcrumb = { title: "Subscribe", path: "/subscribe/" };
 const SEARCH: Breadcrumb = {
   title: "Search",
   path: "/search?status=forecasted,posted",
 };
 export const SUBSCRIBE_CONFIRMATION: Breadcrumb = {
   title: "Confirmation",
-  path: "/newsletter/confirmation/",
+  path: "/subscribe/confirmation/",
 };
 const UNSUBSCRIBE: Breadcrumb = {
   title: "Unsubscribe",
-  path: "/newsletter/unsubscribe/",
+  path: "/subscribe/unsubscribe/",
 };
 
 export const RESEARCH_CRUMBS: BreadcrumbList = [HOME, RESEARCH];
