+++ frontend/tests/e2e/research.spec.ts
@@ -13,11 +13,11 @@ test("has title", async ({ page }) => {
   await expect(page).toHaveTitle("Research | Simpler.Grants.gov");
 });
 
-test("can navigate to /newsletter", async ({ page }) => {
+test("can navigate to /subscribe", async ({ page }) => {
   await page
     .getByRole("link", { name: /sign up for project updates/i })
     .getByTestId("button")
     .click();
 
-  await expect(page).toHaveURL(/newsletter/);
+  await expect(page).toHaveURL(/subscribe/);
 });
