+++ frontend/tests/utils/getRoutes.test.ts
@@ -28,14 +28,14 @@ describe("getNextRoutes", () => {
     expect(result).toEqual([
       "/dev/feature-flags",
       "/health",
-      "/newsletter/confirmation",
-      "/newsletter",
-      "/newsletter/unsubscribe",
       "/opportunity/1",
       "/",
       "/process",
       "/research",
       "/search",
+      "/subscribe/confirmation",
+      "/subscribe",
+      "/subscribe/unsubscribe",
     ]);
   });
 });
@@ -45,10 +45,6 @@ function getPaths() {
     "src/app/[locale]/dev/feature-flags/FeatureFlagsTable.tsx",
     "src/app/[locale]/dev/feature-flags/page.tsx",
     "src/app/[locale]/health/page.tsx",
-    "src/app/[locale]/newsletter/NewsletterForm.tsx",
-    "src/app/[locale]/newsletter/confirmation/page.tsx",
-    "src/app/[locale]/newsletter/page.tsx",
-    "src/app/[locale]/newsletter/unsubscribe/page.tsx",
     "src/app/[locale]/page.tsx",
     "src/app/[locale]/process/ProcessIntro.tsx",
     "src/app/[locale]/process/ProcessInvolved.tsx",
@@ -65,6 +61,10 @@ function getPaths() {
     "src/app/[locale]/search/error.tsx",
     "src/app/[locale]/search/loading.tsx",
     "src/app/[locale]/search/page.tsx",
+    "src/app/[locale]/subscribe/SubscriptionForm.tsx",
+    "src/app/[locale]/subscribe/confirmation/page.tsx",
+    "src/app/[locale]/subscribe/page.tsx",
+    "src/app/[locale]/subscribe/unsubscribe/page.tsx",
     "src/app/api/BaseApi.ts",
     "src/app/api/SearchOpportunityAPI.ts",
     "src/app/api/mock/APIMockResponse.json",
